KisanAI-master/backend/requirements.txt
flask
flask-cors
scikit-learn
joblib
prisma
pyjwt
bcrypt
google-generativeai
python-dotenv
